model_backend/pyproject.toml
[tool.poetry]
name = "translator"
version = "0.1.0"
description = "en-ru\ru-en translation service"
authors = ["Timur Smirnov <smrtmr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ray = {extras = ["serve"], version = "^2.35.0"}
torch = "2.2.2"
transformers = "4.39.3"
sentencepiece = "0.2.0"
opencensus = "0.11.4"
sacremoses = "0.1.1"
pydantic-settings = "^2.2.1"
fastapi = "^0.114.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.0.2"
pytest-asyncio = "0.23.5"
pytest-mock = "3.12.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = "src"
addopts = "-ra -q"
testpaths = [
  "tests"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
